Fix unit_generator links in architecture docs, move resolve_all_features to cargo_compile top-level module
